update projectors unittest
