+++ Wrappers/Python/test/test_PluginsTigre_General.py
@@ -20,6 +20,7 @@ from cil.framework.framework import ImageGeometry
 import numpy as np
 from cil.utilities.display import show2D
 from cil.utilities import dataexample
+from utils_projectors import TestCommon_ProjectionOperatorBlockOperator
 
 from utils import has_tigre, has_nvidia, initialise_tests
 
@@ -313,54 +314,16 @@ class TestMechanics(unittest.TestCase):
         diff = (fp1 - fp2).abs().sum()
         self.assertGreater(diff,0.1)
 
-class TestTIGREBlockOperator(unittest.TestCase):
+class TestTIGREBlockOperator(unittest.TestCase, TestCommon_ProjectionOperatorBlockOperator):
     def setUp(self):
         data = dataexample.SIMULATED_PARALLEL_BEAM_DATA.get()
         self.data = data.get_slice(vertical='centre')
-
-    def test_partition(self):
-        data = self.data.partition(10, 'sequential')
         ig = self.data.geometry.get_ImageGeometry()
-
-        K = ProjectionOperator(image_geometry=ig, acquisition_geometry=data.geometry)
-        A = ProjectionOperator(image_geometry=ig, acquisition_geometry=self.data.geometry)
-
-        u = A.adjoint(self.data)
-        v = K.adjoint(data)
-
-        # the images are not entirely the same as the BlockOperator's requires to 
-        # add all the data of the adjoint operator, which may result in a slightly
-        # different image
-        np.testing.assert_allclose(u.as_array(), v.as_array(), rtol=1.2e-6, atol=1.6e-4)
-
-        x = A.direct(u)
-        y = K.direct(v)
-
-        # let's check that the data is the same
-        k = 0
-        wrong = 0
-        for i, el in enumerate(y.containers):
-            for j in range(el.shape[0]):
-                try:
-                    np.testing.assert_allclose(el.as_array()[j], x.as_array()[k], atol=7e-2, rtol=1e-6)
-                except AssertionError as ae:
-                    print(ae)
-                    wrong += 1
-                # show2D([el.as_array()[j], x.as_array()[k]], cmap=['inferno', 'inferno'])
-                k += 1
-
-        assert wrong == 0
+        self.datasplit = self.data.partition(10, 'sequential')
         
-        # reassemlbe the data
-        out = x * 0
-        k = 0
-        for i, el in enumerate(y.containers):
-            # print (i, el.shape)
-            for j in range(el.shape[0]):
-                out.array[k] = el.as_array()[j]
-                k += 1
 
-        show2D([out, x, out-x], cmap=['inferno', 'inferno', 'seismic'], title=['out', 'x', 'diff'], \
-            num_cols=3)
-        np.testing.assert_allclose(out.as_array(), x.as_array(), atol=1e-2, rtol=1e-6)
+        K = ProjectionOperator(image_geometry=ig, acquisition_geometry=self.datasplit.geometry)
+        A = ProjectionOperator(image_geometry=ig, acquisition_geometry=self.data.geometry)
+        self.projectionOperator = (A, K)
 
+
